refactor: consolidate tool modules into table, record, record_write
- Merge introspection.py + utility.py into table.py (table_describe, table_query, table_aggregate, build_query)
- Merge introspection.py (record_get) + relationships.py into record.py (record_get, rel_references_to, rel_references_from)
- Extract write tools from record.py into record_write.py (7 CRUD/preview/apply tools)
- Remove dev_utils.py entirely (dev_toggle, dev_set_property removed)
- Enrich table_describe with sys_db_object and sys_documentation data
- Update packages.py: add record_write group, rename introspection_only to core_readonly
- Read-only packages (core_readonly, readonly, analyst) exclude record_write
