+++ src/servicenow_mcp/packages.py
@@ -1,18 +1,17 @@
 """Tool package registry and loader for the ServiceNow MCP server."""
 
 _TOOL_GROUP_MODULES: dict[str, str] = {
-    "introspection": "servicenow_mcp.tools.introspection",
-    "relationships": "servicenow_mcp.tools.relationships",
+    "table": "servicenow_mcp.tools.table",
+    "record": "servicenow_mcp.tools.record",
+    "record_write": "servicenow_mcp.tools.record_write",
     "testing": "servicenow_mcp.tools.testing",
     "metadata": "servicenow_mcp.tools.metadata",
     "changes": "servicenow_mcp.tools.changes",
     "debug": "servicenow_mcp.tools.debug",
-    "developer": "servicenow_mcp.tools.developer",
-    "dev_utils": "servicenow_mcp.tools.dev_utils",
     "investigations": "servicenow_mcp.tools.investigations",
     "documentation": "servicenow_mcp.tools.documentation",
-    "utility": "servicenow_mcp.tools.utility",
     "workflow": "servicenow_mcp.tools.workflow",
+    "flow_designer": "servicenow_mcp.tools.flow_designer",
     "domain_incident": "servicenow_mcp.tools.domains.incident",
     "domain_change": "servicenow_mcp.tools.domains.change",
     "domain_cmdb": "servicenow_mcp.tools.domains.cmdb",
@@ -25,25 +24,23 @@ _TOOL_GROUP_MODULES: dict[str, str] = {
 # Registry mapping package names to lists of tool group names.
 # Tool groups correspond to modules in servicenow_mcp.tools.
 PACKAGE_REGISTRY: dict[str, list[str]] = {
-    "introspection_only": [
-        "introspection",
-        "relationships",
+    "core_readonly": [
+        "table",
+        "record",
         "metadata",
-        "utility",
     ],
     "full": [
-        "introspection",
-        "relationships",
+        "table",
+        "record",
+        "record_write",
         # "testing",  # ATF - disabled
         "metadata",
         "changes",
         "debug",
-        "developer",
-        "dev_utils",
         "investigations",
         "documentation",
-        "utility",
         "workflow",
+        "flow_designer",
         "domain_incident",
         "domain_change",
         "domain_cmdb",
@@ -54,92 +51,102 @@ PACKAGE_REGISTRY: dict[str, list[str]] = {
     ],
     "none": [],
     "itil": [
-        "introspection",
-        "relationships",
+        "table",
+        "record",
+        "record_write",
         "metadata",
         "changes",
         "debug",
         "documentation",
-        "utility",
         "workflow",
+        "flow_designer",
         "domain_incident",
         "domain_change",
         "domain_problem",
         "domain_request",
     ],
     "developer": [
-        "introspection",
-        "relationships",
+        "table",
+        "record",
+        "record_write",
         "metadata",
         "changes",
         "debug",
-        "developer",
-        "dev_utils",
         "investigations",
         "documentation",
-        "utility",
         "workflow",
+        "flow_designer",
     ],
     "readonly": [
-        "introspection",
-        "relationships",
+        "table",
+        "record",
         "metadata",
         "changes",
         "debug",
         "investigations",
         "documentation",
-        "utility",
         "workflow",
+        "flow_designer",
     ],
     "analyst": [
-        "introspection",
-        "relationships",
+        "table",
+        "record",
         "metadata",
         "investigations",
         "documentation",
-        "utility",
         "workflow",
+        "flow_designer",
     ],
     "incident_management": [
-        "introspection",
-        "utility",
+        "table",
+        "record",
+        "record_write",
         "domain_incident",
         "debug",
         "workflow",
+        "flow_designer",
     ],
     "change_management": [
-        "introspection",
-        "utility",
+        "table",
+        "record",
+        "record_write",
         "domain_change",
         "changes",
+        "flow_designer",
     ],
     "cmdb": [
-        "introspection",
-        "relationships",
-        "utility",
+        "table",
+        "record",
+        "record_write",
         "domain_cmdb",
     ],
     "problem_management": [
-        "introspection",
-        "utility",
+        "table",
+        "record",
+        "record_write",
         "domain_problem",
         "debug",
         "workflow",
+        "flow_designer",
     ],
     "request_management": [
-        "introspection",
-        "utility",
+        "table",
+        "record",
+        "record_write",
         "domain_request",
         "workflow",
+        "flow_designer",
     ],
     "knowledge_management": [
-        "introspection",
-        "utility",
+        "table",
+        "record",
+        "record_write",
         "domain_knowledge",
     ],
     "service_catalog": [
-        "introspection",
-        "utility",
+        "table",
+        "record",
+        "record_write",
         "domain_service_catalog",
     ],
 }
@@ -168,7 +175,7 @@ def get_package(name: str) -> list[str]:
 
     groups = [g for g in stripped_groups if g]
 
-    if not groups:  # pragma: no cover — defensive guard; line 150 catches all empty strings first
+    if not groups:  # pragma: no cover - defensive guard; line above catches all empty strings first
         raise ValueError("No empty groups allowed")
 
     seen: set[str] = set()
